Reformat project, add configs

+++ src/main/java/com/alan199921/astral/dimensions/innerrealm/InnerRealmDimension.java
@@ -40,21 +40,21 @@ public class InnerRealmDimension extends Dimension {
     @Override
     public float calculateCelestialAngle(long worldTime, float partialTicks) {
 
-            int j = 6000;
-            float f1 = (j + partialTicks) / 24000.0F - 0.25F;
+        int j = 6000;
+        float f1 = (j + partialTicks) / 24000.0F - 0.25F;
 
-            if (f1 < 0.0F) {
-                f1 += 1.0F;
-            }
+        if (f1 < 0.0F) {
+            f1 += 1.0F;
+        }
 
-            if (f1 > 1.0F) {
-                f1 -= 1.0F;
-            }
+        if (f1 > 1.0F) {
+            f1 -= 1.0F;
+        }
 
-            float f2 = f1;
-            f1 = 1.0F - (float) ((Math.cos(f1 * Math.PI) + 1.0D) / 2.0D);
-            f1 = f2 + (f1 - f2) / 3.0F;
-            return f1;
+        float f2 = f1;
+        f1 = 1.0F - (float) ((Math.cos(f1 * Math.PI) + 1.0D) / 2.0D);
+        f1 = f2 + (f1 - f2) / 3.0F;
+        return f1;
     }
 
 
@@ -66,7 +66,7 @@ public class InnerRealmDimension extends Dimension {
     @Override
     @OnlyIn(Dist.CLIENT)
     public Vec3d getFogColor(float p_76562_1_, float p_76562_2_) {
-        float f = MathHelper.cos(p_76562_1_ * ((float)Math.PI * 2F)) * 2.0F + 0.5F;
+        float f = MathHelper.cos(p_76562_1_ * ((float) Math.PI * 2F)) * 2.0F + 0.5F;
         f = MathHelper.clamp(f, 0.0F, 1.0F);
         float f1 = 0.7529412F;
         float f2 = 0.84705883F;
@@ -74,7 +74,7 @@ public class InnerRealmDimension extends Dimension {
         f1 = f1 * (f * 0.94F + 0.06F);
         f2 = f2 * (f * 0.94F + 0.06F);
         f3 = f3 * (f * 0.91F + 0.09F);
-        return new Vec3d((double)f1, (double)f2, (double)f3);
+        return new Vec3d(f1, f2, f3);
     }
 
     @Override
